backend/pyproject.toml
[project]
name = "tododoo-backend"
version = "0.1.0"
description = "FastAPI backend for Tododoo application"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.17.2",
    "fastapi[standard]>=0.121.3",
    "fractional-indexing>=0.1.3",
    "passlib[bcrypt]>=1.7.4",
    "psycopg2-binary>=2.9.11",
    "pwdlib[argon2,bcrypt]>=0.3.0",
    "pyjwt>=2.10.1",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.20",
    "sqlalchemy>=2.0.44",
    "uvicorn[standard]>=0.38.0",
]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "httpx>=0.28.0",
    "pyright>=1.1.330",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --strict-markers"

[tool.pyright]
pythonVersion = "3.12"
typeCheckingMode = "standard"
reportUnusedImport = true
reportUnusedVariable = true
reportGeneralTypeIssues = true
reportOptionalMemberAccess = true
reportOptionalSubscript = true
reportPrivateUsage = false
include = ["app"]
